backend/requirements.txt
fastapi
uvicorn
pydantic
trueskill
pandas
python-multipart
pytest
httpx
scipy
google-cloud-firestore
